#fix #jsx Removed unnecessary form elements and simplified the input fields in the registration form. Refactored form components into InfoForm component.
